+++ RS/src/Model/Traffic/TrafficPatternDefines.cpp
@@ -118,7 +118,7 @@ unsigned int TrafficPatternDefines::size(const std::map<unsigned int, std::strin
 #ifdef DEBUG_POINTS_METHODS
     std::cout << "Model/TrafficPattern/TrafficPatternDefines::size" << std::endl;
 #endif
-    return m.size();
+    return (unsigned int) m.size();
 }
 
 bool TrafficPatternDefines::keyExist(const std::map<unsigned int, std::string> &m, unsigned int key) const {
